refactor: move test setup code to main_test.go and use testConfig struct
